docs: move description not prod ready stories
